requirements.txt
pydantic==1.10.13
pytz
azure-functions
azure-eventgrid
azure-identity
